feat: Add option to override default track from user preferences
(Related to US131040)

Producer X attaches a single <track> element to the Media Player. This
<track> element loads the captions file defined in its "src" attribute,
then passes the loaded captions data to Producer X.

This loading happens when the <track> becomes enabled in Media Player's
settings. When the <track> has a "default" attribute, we automatically
enable that <track> as soon as it's added to the Media Player.

When the current language changes in Producer X, we remove the <track>
and replace it with a new <track> set to a different src and locale.
This new <track> has the "default" attribute, to ensure that it's
enabled and that captions are loaded right away.

However, if the user has a preferred captions language saved in their
user preferences, the "default" attribute is ignored and the track is
not enabled. This prevents captions from being loaded, resulting in an
empty captions list in Producer X.

To get around this issue, this commit adds a new <track> attribute:
default-ignore-preferences. If it is set on a <track>, then user
preferences are ignored and the <track> is always enabled by default.