fix: Sync displayed captions cue text when tracks are loaded
When a track is loaded, it might replace the captions for the currently-
selected captions locale. When this happens, we expect the Media Player
to stop showing the previous captions.

However, this was not the case, and a stale captions cue would linger on
the screen until the player progressed to the timestamp of a new
caption.