US142768: Fix video zooming and positioning for Mixed Layouts slider
With the previous logic, whenever the video enlarged passed a certain
threshold, it would stop enlarging and instead get pushed off to the
side, causing the Media Player to display only a corner of the video.
This seems to be a browser limitation of scaling videos using the
"width" and "height" properties.

The new logic uses CSS' transform(), which allows us to scale up the
video without limitations, as well as simplify our zooming logic.