US142768: Fix accuracy & aspect ratio handling of Mixed Layout slider
The calculations for the Mixed Layout slider now use the unscaled
dimensions of the loaded video, making the calculations more accurate
and easier to understand.

In addition, the calculations now have a fix for layout frames whose
aspect ratio is bigger height-wise compared to the aspect ratio of the
full video. In this scenario, we limit the scaling based on height
rather than by width.